backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.9.0
google-genai>=0.2.0
python-dotenv>=1.0.1
sse-starlette>=2.1.0
langchain>=0.2.0
langchain-google-genai>=1.0.0
langchain-community>=0.2.0
chromadb>=0.5.0
pypdf>=4.0.0
langgraph>=0.1.0
pandas>=2.0.0
